refactor (Help command): Extract help utilities to their own file
